refactored code so that we can update routes in the appropriate py file
