--- flask_app/.gitignore
@@ -1,1 +1,0 @@
-.app.yaml
