+++ flask_app/__init__.py
@@ -1,5 +1,2 @@
 #This file initializes your application and brings together all of the various components.
 #
-from flask import Flask
-
-app = Flask(__name__)
